rename app --> tui, move ui to tui
