+++ src/lib.rs
@@ -3,4 +3,3 @@ pub mod args;
 pub mod cli;
 pub mod execution;
 pub mod telemetry;
-pub mod tui;
